Refactor command handling to use MediatR
This commit refactors the command handling mechanism to utilize MediatR for sending commands, enhancing decoupling and aligning with the Mediator pattern.

Key changes include:
- Updated `ProductEndpoints.cs` and `UserEndpoints.cs` to use `IMediator` instead of specific command handler types.
- Moved `ICommandHandler` interface to the `DigitalRightsManagement.Application.Messaging` namespace for better organization.
- Reorganized repository interfaces (`IProductRepository`, `IUnitOfWork`, `IUserRepository`) into the `DigitalRightsManagement.Application.Persistence` namespace.
- Ensured consistent use of `CancellationToken` in command handlers.
- Added necessary using directives to accommodate the new structure.

These changes improve code organization and promote better practices in command handling and data access.
