+++ src/DigitalRightsManagement.Application/Messaging/ICommandHandler.cs
@@ -1,7 +1,7 @@
 ﻿using DigitalRightsManagement.Common.Messaging;
 using MediatR;
 
-namespace DigitalRightsManagement.Application;
+namespace DigitalRightsManagement.Application.Messaging;
 
 public interface ICommandHandler<in TCommand, TResponse> : IRequestHandler<TCommand, TResponse>
     where TCommand : ICommand<TResponse>;
