+++ src/DigitalRightsManagement.Application/Persistence/IUnitOfWork.cs
@@ -1,4 +1,4 @@
-﻿namespace DigitalRightsManagement.Application;
+﻿namespace DigitalRightsManagement.Application.Persistence;
 
 public interface IUnitOfWork
 {
