+++ src/DigitalRightsManagement.Application/ProductAggregate/CreateProduct.cs
@@ -1,19 +1,21 @@
 ﻿using Ardalis.Result;
-using DigitalRightsManagement.Domain.ProductAggregate;
+using DigitalRightsManagement.Application.Messaging;
+using DigitalRightsManagement.Application.Persistence;
 using DigitalRightsManagement.Common.DDD;
 using DigitalRightsManagement.Common.Messaging;
+using DigitalRightsManagement.Domain.ProductAggregate;
 
 namespace DigitalRightsManagement.Application.ProductAggregate;
 
 public sealed class CreateProductCommandHandler(IUserRepository userRepository, IProductRepository productRepository, IUnitOfWork unitOfWork) : ICommandHandler<CreateProductCommand, Result>
 {
-    public async Task<Result> Handle(CreateProductCommand command, CancellationToken ct)
+    public async Task<Result> Handle(CreateProductCommand command, CancellationToken cancellationToken)
     {
-        return await userRepository.GetById(command.UserId, ct)
+        return await userRepository.GetById(command.UserId, cancellationToken)
             .DoubleBind(user => Price.Create(command.Price, command.Currency))
             .BindAsync(t => Product.Create(command.Name, command.Description, t.Next, t.Prev.Id))
             .Tap(productRepository.Add)
-            .Tap(_ => unitOfWork.SaveChanges(ct))
+            .Tap(_ => unitOfWork.SaveChanges(cancellationToken))
             .MapAsync(_ => Result.Success());
     }
 }
