+++ src/DigitalRightsManagement.Application/UserAggregate/ChangeUserRole.cs
@@ -1,4 +1,6 @@
 ﻿using Ardalis.Result;
+using DigitalRightsManagement.Application.Messaging;
+using DigitalRightsManagement.Application.Persistence;
 using DigitalRightsManagement.Common.DDD;
 using DigitalRightsManagement.Common.Messaging;
 using DigitalRightsManagement.Domain.UserAggregate;
